Use Proxy queue + Python queue for Python UDF flow control credits (#2206)
This PR makes the Python flow control credits to be used by the flow
control component.
- The proxy side will keep track of credit from PVM's internal queue.
- The flow control will use the credit from the Proxy's queue and credit
from PVM's internal queue together as the credit of a Python Worker.
- We calculate the data message sizes with library `pympler`.

---------

+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/common/AkkaMessageTransferService.scala
@@ -90,7 +90,7 @@ class AkkaMessageTransferService(
     }
   }
 
-  def updateChannelCreditFromReceiver(channel: ChannelID, credit: Int): Unit = {
+  def updateChannelCreditFromReceiver(channel: ChannelID, credit: Long): Unit = {
     val flowControl = channelToFC.getOrElseUpdate(channel, new FlowControl())
     flowControl.isPollingForCredit = false
     flowControl.updateCredit(credit)
