+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/worker/WorkflowWorker.scala
@@ -86,7 +86,7 @@ class WorkflowWorker(
   }
 
   /** flow-control */
-  override def getSenderCredits(channelID: ChannelID): Int =
+  override def getSenderCredits(channelID: ChannelID): Long =
     dp.getSenderCredits(channelID)
 
   override def initState(): Unit = {
