+++ core/amber/src/test/scala/edu/uci/ics/amber/engine/architecture/control/utils/TrivialControlTester.scala
@@ -32,7 +32,7 @@ class TrivialControlTester(
   }
 
   /** flow-control */
-  override def getSenderCredits(channelID: ChannelID): Int = ???
+  override def getSenderCredits(channelID: ChannelID): Long = ???
 
   override def initState(): Unit = {}
